Updated notification functionality

+++ src/java/com/hakim/dao/NotificationsDao.java
@@ -2,6 +2,8 @@ package com.hakim.dao;
 
 import java.sql.Connection;
 import java.sql.PreparedStatement;
+import java.sql.ResultSet;
+import java.sql.SQLException;
 
 /**
  *
@@ -14,12 +16,13 @@ public class NotificationsDao {
         this.con = con;
     }
     
-    public boolean addNotification(int pid,String message){
-        String query="insert into notifications(message,pid) value(?,?)";
+    public boolean addNotification(int pid,String message,int uid){
+        String query="insert into notifications(message,pid,uid) value(?,?,?)";
         boolean added=false;
         try(PreparedStatement statement=con.prepareStatement(query)){
             statement.setString(1, message);
             statement.setInt(2, pid);
+            statement.setInt(3, uid);
             
             statement.execute();
             added=true;
@@ -31,11 +34,12 @@ public class NotificationsDao {
         return added;
     }
     
-    public boolean removeNotification(int nid){
-        String query="delete from notifications where nid=?";
+    public boolean removeNotification(int nid,int uid){
+        String query="delete from notifications where nid=? & uid";
         boolean removed=false;
         try(PreparedStatement statement=con.prepareStatement(query)){
             statement.setInt(1, nid);
+            statement.setInt(2, uid);
             
             
             statement.executeUpdate();
@@ -46,5 +50,25 @@ public class NotificationsDao {
         
         
         return removed;
+    }
+    
+    public int count(int uid){
+        int count=0;
+        
+        String query="select count(*) from notifications where uid = ?";
+        try(PreparedStatement statement=con.prepareStatement(query)){
+            statement.setInt(1, uid);
+           
+            
+            ResultSet set=statement.executeQuery();
+            if(set.next()){
+                count=set.getInt(1);
+            }
+            
+        }catch(SQLException e){
+            e.printStackTrace();
+        }
+        
+        return count;
     }
 }
